Organize form input components
fixes #67
