+++ app/listings/create/page.tsx
@@ -2,9 +2,9 @@
 
 import Button from '@/components/button';
 import FormInput from '@/components/form-input';
-import FormInputMap from '@/components/form-input-map';
-import FormInputTextArea from '@/components/form-input-textarea';
-import FormReset from '@/components/form-reset';
+import FormInputMap from '@/components/form-input/map';
+import FormInputReset from '@/components/form-input/reset';
+import FormInputTextArea from '@/components/form-input/textarea';
 import TextError from '@/components/text/error';
 import { fetchDateOneYearFromToday, fetchToday } from "@/lib/date";
 import { useFormState } from 'react-dom';
@@ -57,7 +57,7 @@ export default function CreateListing() {
                         </div>
                     </div>
                     <div className="flex justify-end space-x-8 pt-8">
-                        <FormReset />
+                        <FormInputReset />
                         <Button text='Submit' />
                     </div>
                 </div>
