+++ components/form-input/reset.tsx
@@ -1,4 +1,4 @@
-export default function FormReset() {
+export default function FormInputReset() {
     return <input type="reset" value="Reset"
         className="cursor-pointer rounded-md border-2 border-transparent px-6 py-3 transition-all hover:border-slate-200" />;
 }
