+++ stories/form-input/FormInputTextArea.stories.tsx
@@ -1,4 +1,4 @@
-import FormInputTextArea from "@/components/form-input-textarea";
+import FormInputTextArea from "@/components/form-input/textarea";
 import type { Meta, StoryObj } from "@storybook/react";
 
 const meta: Meta<typeof FormInputTextArea> = {
